fix: vuetify nuxt plugin cannot resolve imports
